+++ tmpl/default_pro_banner_options.php
@@ -9,22 +9,15 @@
 // No direct access.
 \defined('_JEXEC') or die;
 
-use Joomla\CMS\Factory;
 use Joomla\CMS\Language\Text;
 
-Factory::getApplication()->getDocument()->getWebAssetManager()
-	->usePreset('choicesjs')
-	->useScript('webcomponent.field-fancy-select');
 ?>
 
-<div class="row">
-	<div class="collapse resethits-options" id="pro-banner-options" data-bs-parent="#resethits">
-		<hr/>
-		<joomla-alert type="danger" class="joomla-alert--show" role="alert">
-			<div class="alert-heading"><span class="danger"></span><span class="visually-hidden">Error</span></div>
-			<div class="alert-wrapper">
-				<div class="alert-message"><?php echo Text::_('MOD_RESETHITS_PRO_ONLY'); ?></div>
-			</div>
-		</joomla-alert>
+	<div class="row">
+		<div class="collapse resethits-options" id="pro-banner-options" data-bs-parent="#resethits">
+			<hr/>
+			<a class="btn btn-success btn-sm resethits-pro-badge" href="https://www.joomill-extensions.com/extensions/reset-article-views-hits-counter" target="_blank" rel="noopener noreferrer">
+				<span class="icon-star icon-white" aria-hidden="true"></span> <?php echo Text::_('MOD_RESETHITS_PRO_ONLY'); ?>
+			</a>
+		</div>
 	</div>
-</div>